// src/app/api/models/page-bc-resources-it-infrastructure.ts
/* tslint:disable */
/* eslint-disable */
import { BcResourcesItInfrastructure } from './bc-resources-it-infrastructure';
import { PageableObject } from './pageable-object';
import { SortObject } from './sort-object';
export interface PageBcResourcesItInfrastructure {
  content?: Array<BcResourcesItInfrastructure>;
  empty?: boolean;
  first?: boolean;
  last?: boolean;
  number?: number;
  numberOfElements?: number;
  pageable?: PageableObject;
  size?: number;
  sort?: SortObject;
  totalElements?: number;
  totalPages?: number;
}