// src/app/api/models/bc-resources-it-infrastructure.ts
/* tslint:disable */
/* eslint-disable */
import { BcResources } from './bc-resources';
export interface BcResourcesItInfrastructure {
  availableCount?: number;
  detailsAr?: string;
  detailsEn?: string;
  id?: number;
  isActive?: boolean;
  purchasedCount?: number;
  requiredCount?: number;
  resource: BcResources;
  specialInstruction?: string;
}